requirements.txt
google
python-dotenv
markdownify
beautifulsoup4
fastapi
asyncio
re
subprocess
os
manim
requests